Created pointers_overview.c, pointers_overview_continued.c on this branch accidentally, will move them to main
