+++ Pointers/Pointer_Basics/pointers_overview.c
@@ -2,130 +2,53 @@
 #include <stdlib.h>
 #include <limits.h>
 
-#define MAX_STACK_SIZE 10
 
 /**
- * Implementation of an `int` stack data structure.
+ * @brief Tries to swap a,b. Will not work,
+ *        because the FUNCTION's a,b are LOCAL
+ *        to it, and don't have the same ADDRESSES
+ *        as the a,b passed to it. therefore, it will
+ *        swap its own a,b, but not affect what was passed
+ *        to it. 
  * 
- * @param stack_arr An `int` array, which will hold the stack's data
- * @param size The maximum size of the stack
- * @param top Pointer to the top of the stack
-*/
-typedef struct stack
+ * @param first First number
+ * @param second Second number
+ */
+void local_int_swap_wrong(int first, int second)
 {
-    int* stack_arr; //An array of `int` that implements our stack's data storage
-    int size; //Size of `stack_arr`
-    int* top; //A POINTER to the top of the stack
-} Stack;
-
-/**
- * Creates an `int` stack with a specified maximum size
- * 
- * @param size The maximum size of the stack
- * 
- * @returns A new stack with the given maximum size,
- *          or NULL on error.
-*/
-Stack* create_stack(int size)
-{
-    if(size <= 0) return 0;
-
-    Stack* stack = (Stack*)calloc(1, sizeof(Stack)); //In order to use the dynamic allocation functions, we MUST use a pointer!
-    if(!stack) return 0;
-
-    stack->stack_arr = (int*)calloc(size, sizeof(int));
-    if(!(stack->stack_arr))
-    {
-        free(stack);
-        return 0;
-    }
-    stack->size = size;
-    stack->top = stack->stack_arr;
-
-    return stack;
+    int temp = second;
+    second = first;
+    first = temp;    
 }
 
 /**
- * Checks if the given stack is empty.
+ * @brief Swaps the values of two `int` variables inplace.
+ *        This is the correct way to do this kind of inplace swapping,
+ *        since we can directly access the VALUES at the ADDRESSES
+ *        passed to the function.
  * 
- * @param stack Given stack
- * 
- * @returns 1 if the stack is empty, 0 if not - or on error.
-*/
-int is_empty(Stack* stack)
+ * @param first First number
+ * @param second Second number
+ */
+void local_int_swap_correct(int* first, int* second)
 {
-    if(!stack || !(stack->stack_arr) || (stack->top == stack->stack_arr)) return 1;
-    else return 0;
+    //Notice that since we need to access the VALUES
+    //at the POINTERS (addresses) `first` and `second`
+    //using the DEREFERENCE OPERATOR, we MUST first make
+    //sure that the pointers themselves are not NULL.
+    //The dereference operation on a NULL pointer is UNDEFINED,
+    //and will cause the program to crash.
+    if(!first || !second) return;
+
+    int temp = *second;
+    *second = *first;
+    *first = temp;
 }
 
-/**
- * Checks if the given stack is full
- * 
- * @param stack Given stack
- * 
- * @returns 1 if the stack is full, 0 if not - or on error.
-*/
-int is_full(Stack* stack)
-{
-    if(!stack || !(stack->stack_arr)) return 0;
-    else if((stack->top) - (stack->stack_arr) == (stack->size )) return 1;
-
-    return 0;
-}
-
-/**
- * Pushes a given value to a given stack
- * 
- * @param stack Given stack
- * @param value Value to push onto the stack
- * 
- * @returns 1 if the value was pushed successfully, 0 if
- *          the stack is full or on error.
-*/
-int push(Stack* stack, int value)
-{
-    if(!stack || !(stack->stack_arr) || is_full(stack)) 
-        return 0; //Checking that the stack exists, and is not full.
-
-    stack->top++; //This is pointer arithmetic, we moved `top` to point to the next address in `stack_arr`
-    *(stack->top) = value; //Here, the left hand side of the '=' is an ACCESS TO THE VALUE OF A POINTER
-    
-    return 1;
-}
-
-/**
- * Pops the element at the top of the stack, and returns its value.
- * 
- * @param stack Given stack
- * 
- * @returns The element that was at the top of the stack,
- *          or `INT_MAX` if the stack is empty, or on error.
-*/
-int pop(Stack* stack)
-{
-    if(!stack || !(stack->top) || is_empty(stack)) return INT_MAX;
-
-    int top = *(stack->top); //Here, the right hand side of the '=' is an ACCESS TO THE VALUE OF A POINTER
-    stack->top--; //This is pointer arithmetic, we moved `top` to point to the previous address in `stack_arr`
-
-    return top;
-}
-
-void destroy_stack(Stack* stack)
-{
-    if(!stack) return;
-
-    if(stack->stack_arr)
-    {
-        free(stack->stack_arr);
-        stack->stack_arr = 0;
-    }
-
-    free(stack);
-}
 
 int main(void)
 {
+    printf("*********************************POINTERS - OVERVIEW:*********************************\n");
     //We'll explore the concept of pointers in C,
     //their syntax and uses.
 
@@ -137,6 +60,7 @@ int main(void)
     //OR
     //<type name> *<variable name>;
 
+    printf("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~POINTERS - SYNTAX:~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n");
     //For example:
     int* int_ptr;
     float *float_ptr;
@@ -161,31 +85,55 @@ int main(void)
     //When coding in C, pointers are our way of keeping track of memory addresses, and
     //determining what data is stored in them.
 
-    //In order to dynamically allocate memory, the use of pointers is MANDATORY.
-    //Let's take a look at an example implementation of a stack data structure (defined above):
-    Stack* stack = create_stack(MAX_STACK_SIZE); //We have to use a POINTER to a `Stack`!
+    printf("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~POINTERS - SIMPLE SWAP FUNCTION:~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n");
+    //The uses for pointers are many and varied, but let's see
+    //a very basic example;
+    //Say we have two variables (possibly elements in an array),
+    //and we wish to swap their values. We can implement this basic and common operation
+    //each time we wish to perform it, however, that is tiresome and unnecessary - we would
+    //like to write a function that does that.
+    //Let's take a look:
+    int first = -20;
+    int second = 5;
+    printf("The values of `first`, `second`:\n"
+           "`first`:\t%d\n"
+           "`second`:\t%d\n", first, second);
 
-    if(!stack) return 0; //Checking if `stack` was created successfully
+    //Let's try a naive approach. Please see the function `void local_int_swap_wrong(int first, int second)`
+    //defined above:
+    local_int_swap_wrong(first, second);
+    printf("The values of `first`, `second`:\n"
+           "`first`:\t%d\n"
+           "`second`:\t%d\n", first, second);
 
-    int i = 0;
+    //Nothing happened! Why is that?
 
-    //Pushing the numbers 0..9 onto the stack...
-    while(!(is_full(stack)))
-    {
-        push(stack, i);
-        i++;
-        if(is_full(stack)) printf("Stack is full!\n");
-    }
+    //We need to remember that `first` and `second` are LOCAL variables,
+    //and exist on our `main` function's STACK FRAME.
 
-    printf("Contents of the stack are:\n");
-    while(!(is_empty(stack)))
-    {
-        printf("%d\n", pop(stack));
-        if(is_empty(stack)) printf("Stack is empty!\n");
-    }
+    //When we pass a variable to a function, the VALUE of that variable is COPIED
+    //to the STACK FRAME of the function being called, and assigned to a variable
+    //with the name matching the name of the matching argument in the function's SIGNATURE.
 
-    destroy_stack(stack);
-    stack = 0; //Assigning `0` to `stack` to prevent a "dangling pointer"
+    //In short, the functions we call know NOTHING about the actual VARIABLES we pass to them
+    //as arguments, except for their VALUES, which get COPIED into the function's STACK FRAMES.
+
+    //What we would like to do is pass the ADDRESSES of the two variables whos values we
+    //wish to swap - and that way the function that we're calling can access the
+    //MEMORY CONTENTS AT THE ADDRESSES that we passed. The function we call can
+    //change the contents of the program's memory, and those changes will persist in the
+    //locations whose contents were changed.
+
+    //A CORRECT version of the swap function we wanted is implemented above; Let's try it:
+    int* first_ptr = &first;
+    int* second_ptr = &second;
+    local_int_swap_correct(first_ptr, second_ptr);
+    printf("The values of `first`, `second`:\n"
+           "`first`:\t%d\n"
+           "`second`:\t%d\n", first, second);
+
+    //We used direct memory access to change the values of variables,
+    //and that direct memory access is what pointers provide.
 
     return 0;
 }
